PERCEPTION: separate and organize cpp perception code
